requirements.txt
python-dotenv
langchain
langchain_core
langchain_openai